mount /usr and all dirs

+++ installers/slurm-dev/radiasoft-download.sh
@@ -19,10 +19,9 @@ radia_run slurm-dev
     fi
     install_yum update
     slurm_dev_nfs
-    # this will cause a reboot so do after NFS (which will modify fstab)
-    install_repo_eval redhat-docker
     install_yum slurm-slurmd slurm-slurmctld
-    dd if=/dev/urandom bs=1 count=1024 | install -m 400 -o munge -g munge /dev/stdin /etc/munge/munge.key
+    dd if=/dev/urandom bs=1 count=1024 \
+        | install -m 400 -o munge -g munge /dev/stdin /etc/munge/munge.key
     local f
     for f in munge slurmctld slurmd; do
         systemctl start "$f"
@@ -32,6 +31,9 @@ radia_run slurm-dev
 }
 
 slurm_dev_nfs() {
+    if grep -s -q $_slurm_dev_nfs_server:/usr /etc/fstab; then
+        return
+    fi
     install_yum install nfs-utils
     if ! showmount -e "$_slurm_dev_nfs_server" >&/dev/null; then
         install_error '
@@ -41,6 +43,9 @@ dnf install -y nfs-utils
 cat << EOF > /etc/exports.d/home_vagrant.exports
 /home/vagrant 10.10.10.0/24(rw,root_squash,no_subtree_check,async,secure)
 EOF
+cat << EOF > /etc/exports.d/usr.exports
+/usr 10.10.10.0/24(rw,root_squash,no_subtree_check,async,secure)
+EOF
 systemctl enable nfs-server
 systemctl restart nfs-server
 
@@ -48,10 +53,16 @@ systemctl restart nfs-server
     fi
     # do this first, because we want to mount /etc/fstab on reboot
     local f
-    for f in .local .pyenv src; do
+    for f in ~/.local ~/.pyenv ~/src ~/bin /usr; do
+        mkdir -p "$f"
         echo "$_slurm_dev_nfs_server:$f $f nfs defaults,vers=4.1,soft,noacl,_netdev 0 0"
     done | sudo tee -a /etc/fstab > /dev/null
-
+    install_err 'You need to:
+exit
+vagrant reload
+vssh
+radia_run slurm-dev
+'
 }
 
 slurm_dev_no_nfs() {
